development done

+++ src/service/stock.service.ts
@@ -1,32 +1,47 @@
-import { AppDataSource } from "../data-source";
-import { Stock } from "../entities/stock.entity";
-import { IStockCreate } from "../interfaces/stock";
+import { AppDataSource } from '../data-source';
+import { Stock } from '../entities/stock.entity';
+import { Supply } from '../entities/supply.entity';
+import { AppError } from '../errors/AppError';
+import { IStockCreate } from '../interfaces/stock';
 
 class StockService {
   static create = async ({ qtd, supply_id }: IStockCreate) => {
     const userRepository = AppDataSource.getRepository(Stock);
+    const supplyRepository = AppDataSource.getRepository(Supply);
 
     const stockList = await userRepository.find();
+    const supply = await supplyRepository.findOne({ where: { id: supply_id } });
 
     const supplyAlreadyExist = stockList.find(
-      (stock) => stock.supply_id.id === supply_id
+      (stock) => stock.supply.id === supply_id
     );
 
     if (supplyAlreadyExist) {
-      throw new Error("This item already exist"); // Usar o atualizar
+      const actualQtd = supplyAlreadyExist.qtd;
+      const incomingQtd = qtd;
+
+      const newTotal = actualQtd + incomingQtd;
+
+      await this.update(supplyAlreadyExist.id, newTotal);
+
+      return { message: 'Stock updated' };
     }
 
-    const stock = new Stock();
-    stock.qtd = qtd;
+    if (supply) {
+      const stock = new Stock();
+      stock.qtd = qtd;
+      stock.supply = supply;
 
-    userRepository.create(stock);
-    await userRepository.save(stock);
+      userRepository.create(stock);
+      await userRepository.save(stock);
 
-    return stock;
+      return stock;
+    } 
   };
 
   static list = async () => {
     const userRepository = AppDataSource.getRepository(Stock);
+
     const stockList = await userRepository.find();
 
     return stockList;
@@ -34,29 +49,40 @@ class StockService {
 
   static readOne = async (id: string) => {
     const userRepository = AppDataSource.getRepository(Stock);
+
     const stockList = await userRepository.find();
 
     const stock = stockList.find((stock) => stock.id === id);
+
+    if(!stock){
+      throw new AppError(404, 'Stock not found')
+    }
 
     return stock;
   };
 
   static update = async (id: string, qtd: number) => {
     const userRepository = AppDataSource.getRepository(Stock);
-    const stockList = await userRepository.find();
 
-    const stock = stockList.find((stock) => stock.id === id);
+    const stock = await userRepository.findOne({ where: { id } });
 
+    if (!stock) {
+      throw new AppError(404, 'Stock not found');
+    }
     await userRepository.update(stock!.id, { qtd });
-
     return true;
   };
 
   static delete = async (id: string) => {
     const userRepository = AppDataSource.getRepository(Stock);
+
     const stockList = await userRepository.find();
 
     const stock = stockList.find((stock) => stock.id === id);
+
+    if(!stock){
+      throw new AppError(404, 'Stock not found' )
+    }
 
     await userRepository.delete(stock!.id);
 
@@ -64,4 +90,4 @@ class StockService {
   };
 }
 
-export default StockService
+export default StockService;
